Fix aspect ratio, refactor draw calls, and update cleanup
- Correct aspect ratio calculation in camera perspective matrix.
- Rename vertex shader scale uniform to ballscale.
- Replace per-ball draw calls with model.Draw() in main loop.
- Update cleanup to delete shader program instead of VAO/VBO.

+++ physics-engine/Camera.cpp
@@ -12,7 +12,7 @@ void Camera::updateMatrix(float FOVdeg, float nearPlane, float farPlane) {
 
 
 	view = glm::lookAt(Position , Position + Orientation, Up);
-	projection = glm::perspective(glm::radians(FOVdeg), (float)(width / height), nearPlane, farPlane);
+	projection = glm::perspective(glm::radians(FOVdeg), (float)width / (float)height, nearPlane, farPlane);
 	camMatrix = projection * view;
 }
